test1/pyproject.toml
[tool.poetry]
name = "test1"
version = "0.1.0"
description = "test1"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_covid19_news", "nonebot_plugin_epicfree", "nonebot_plugin_picsearcher", "nonebot_plugin_lolmatch", "nonebot_plugin_htmlrender", "nonebot_plugin_apscheduler", "nonebot_plugin_heweather"]
plugin_dirs = ["test1/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
